feat: createCache interface, doc, test
