+++ src/storage/interface.js
