+++ src/symbol.js
