+++ test/ttl.test.js
@@ -1,9 +1,9 @@
 'use strict'
 
 const t = require('tap')
-const { Cache } = require('..')
 const { promisify } = require('util')
-const createStorage = require('../storage')
+const { Cache } = require('../src/cache')
+const createStorage = require('../src/storage')
 
 const sleep = promisify(setTimeout)
